add dummy data storage file
