+++ src/components/SearchResults/SearchResults.jsx
@@ -9,84 +9,7 @@ import { useEffect, useState } from 'react';
 
 import './SearchResults.css';
 
-const Book = (
-  title,
-  author,
-  publicationDate,
-  imageLink,
-  form,
-  language,
-  isReleased,
-  isReadableOnline,
-  id = ''
-) => {
-  return {
-    title: title,
-    author: author,
-    form: form,
-    publicationDate: publicationDate,
-    imageLink: imageLink,
-    language: language,
-    isReleased: isReleased,
-    isReadableOnline: isReadableOnline,
-    id: id,
-  };
-};
-
-const books = [
-  new Book(
-    'In Search of Lost time',
-    'Marcel Proust',
-    2005,
-    '',
-    'Printed',
-    'English',
-    'true',
-    'false'
-  ),
-  new Book(
-    'Ulysses',
-    'James Joyse',
-    1950,
-    '',
-    'Printed',
-    'French',
-    'false',
-    'false'
-  ),
-  new Book(
-    'Don Quixote',
-    'Muigel de Cervantes',
-    1873,
-    '',
-    'Electronic',
-    'Latvian',
-    'true',
-    'true'
-  ),
-  new Book(
-    'One Hundred Years of Solitude',
-    'Gabriel Garcia Marquez',
-    1701,
-    '',
-    'Printed',
-    'Lithuanian',
-    'false',
-    'true'
-  ),
-  new Book(
-    'The Great Gatsby',
-    'F. Scott Fitzgerald',
-    1992,
-    '',
-    'Electronic',
-    'English',
-    'true',
-    'false'
-  ),
-];
-
-const SearchResults = ({ searchValue, isOnlineReadable = 'false' }) => {
+const SearchResults = ({ searchValue, isOnlineReadable = 'false', books }) => {
   const [bookList, setbookList] = useState([...books]);
   const [searchFilters, setSearchFilters] = useState({
     addressValue: '',
