+++ src/pages/Account/components/AccountBooks/AccountBooks.jsx
@@ -1,0 +1,7 @@
+import './AccountBooks.css';
+
+const AccountBooks = () => {
+  return <div>AccountBooks</div>;
+};
+
+export default AccountBooks;
