+++ src/pages/Account/pages/Overview/Overview.css
@@ -11,6 +11,11 @@
   align-items: center;
 }
 
+.accountBooks {
+  width: 65%;
+  border: 1px solid rgba(0, 0, 0, 0.05);
+}
+
 #avatar {
   width: 100px;
   height: 100px;
